UPdate properties of player Details

+++ Roulette.BusinessLogic/DTO/Requests/PlaceBetRequest.cs
@@ -10,7 +10,7 @@ namespace Roulette.BusinessLogic.DTO.Requests
     {
         public string GameId { get; set; }
         public string Reference { get; set; }
-        public long PlayerId { get; set; }
+        public int PlayerId { get; set; }
         public double StakeAmount { get; set; }
     }
 }
